RSS: fix center() bugs, conversion bugs
The To of an RSS is the minimum corner of the rectangle, not the
center. Fix center() to add half the rectangle dimensions along the
axes to the center.  Fix areas of conversion code where To was
directly being considered the center for RSS.  Add a function to set
the To of an RSS from its center.

Also fix a few bugs in converting to RSS where the computed RSS did
not contain the entire original BV.