octree: use log odds for free/occupied
Previously, the probability was used as a threshold for when FCL
considered a node free or occupied. However, the default was a
probability of 0.0, which is impossible to represent in an octree
(since it stores the logodds). So, instead, have a log-odds threshold
(where 0.0 logodds indicates 0.5 probability).

For more info on log-odds, read the paper "OctoMap: An Efficient
Probabilistic 3D Mapping Framework Based on Octrees"