math/bv: add distanceBV to get min distance
Add a distanceBV utility function to find the lower bound of the minimum
distance between two BVs.

For many BVs, this is not very efficient, as the BVs are simply
converted to AABBs to find the distance. For the various RSS BVs,
though, a good lower bound of the minimum distance between the bounded
geometry will be efficiently returned.