octree_solver performance improvements
Always recurse into octree before BVH. This avoids scenarios where much
time is wasted in parallel surfaces that are not near the minimum.

Assume identity configuration of the octree in the distance and intersect
recursive functions. The interface functions will multiply the
transforms to make this true. This halves the number of transformation
operations that have to be performed during tree traversal.

For distance queries, use an optimal RSS representation of the octomap
cubes to get a very tight lower-bound for the minimum distance. This
greatly enhances the pruning of the distance check.

For collision queries, save some time by creating an OBB directly from
the AABB of the octomap cube. This saves a multiplication by the
identity transform matrix.