add region of interest to octree
Add a set of halfspaces defining a region of interest to octree.
Only parts of the tree on or in all the halfspaces in the region of
interest will be considered for distance queries.

Eventually, this region of interest needs to apply to all octree
operations, but only distance is implemented to start with.