change distance result brief information to intptr_t
The distance result brief information stored the brief information
in an `int`. However, some primatives can only be stored by pointer
(namely octree nodes), and `int` is not wide enough on 64-bit
platforms to safely store a pointer. So upgrade the brief
information to be `intptr_t` which is guaranteed to be wide enough.